requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]
python-multipart

# Database & ORM
sqlmodel
psycopg2-binary
alembic

# Authentication & Security
pyjwt[crypto]
bcrypt
cryptography
slowapi

# Validation
email-validator

# Configuration
python-dotenv
pydantic-settings